+++ CHANGELOG.md
@@ -12,6 +12,11 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.33.0] - 2020-12-02
+### Added
+- Method `get_input_data` on the `FunctionSchedule`-class. 
+- Method `get_input_data` on the `FunctionScheduleAPI`-class.
+
 ## [0.32.0] - 2020-12-03
 ### Fixed
 - Fix the client initialization to correctly pass all keyword arguments.